data-transmission/requirements.txt
certifi==2018.4.16
chardet==3.0.4
happybase==1.1.0
idna==2.7
kafka-python==1.4.3
ply==3.11
requests==2.20.0
schedule==0.5.0
six==1.11.0
thriftpy==0.3.9
urllib3==1.23
websocket-client==0.54.0
